requirements.txt
# TripForge - AI Travel Planner Dependencies

# Core framework
streamlit>=1.28.0

# AI and LLM
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
langchain-core>=0.1.0

# API integration
google-search-results>=2.4.2

# Environment management
python-dotenv>=1.0.0

# Data handling
typing-extensions>=4.8.0

# Optional: Development dependencies
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0